movies/sqlite_to_postgres/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "тест из обучения"
authors = ["Борис Гельруд <agu-66@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.1.1"
python-dotenv = "^1.0.0"
flake8 = "^6.1.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 79
